sending finish signals as broadcasts. NOT WORKING!
Instead of using on-device files to signal that a screenshot was taken,
broadcasts will be sent via the adb connection. This won't work, since
the BroadcastReceiver runs on the main thread. Trying to get the
receiver run on another handler/Looper didn't change anything. The main
thread is blocking until timeout and the broadcast can't be processed
then because the receiver will be already unregistered by then. Well ...
just saving what have been done for later.